Changes Int to Int64(CLong equivalent) for fields that expect unix timestamp

+++ FitpaySDK/Rest/Models/Commit.swift
@@ -4,7 +4,7 @@ public class Commit
     var links:[ResourceLink]?
     var commitType:CommitType?
     var payload:Payload?
-    var created:Int?
+    var created:Int64?
     var previousCommit:String?
     var commit:String?
 }
